pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
  {name = "rojasdaniel", email = "danielmrp93@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"

dependencies = [
  "langgraph-cli[inmem]>=0.2.8,<0.3.0",
  "langgraph-prebuilt==0.1.8",
  "langchain-core>=0.3.58",
  "langchain-community>=0.3.23",
  "langchain-openai>=0.3.16",
  "openai>=1.77.0",
  "python-dotenv>=1.1.0",
  "nltk>=3.9.1",
  "pandas>=2.2.3",
  "streamlit>=1.45.0",
]